NPT-33: Fixed issue where even if the user didn't exist, the delete request was still considered as a success

+++ src/main/java/com/npe/pet/uncaughtHeroes/service/HeroService.java
@@ -61,8 +61,13 @@ public class HeroService {
     }
 
     public void deleteById(String id) {
+        Optional<Hero> heroOptional = heroRepository.findById(id);
+        if (heroOptional.isEmpty()) {
+            throw new HeroNotFoundException("Hero not found with ID: " + id);
+        }
         heroRepository.deleteById(id);
     }
+
 
     public List<Hero> findByStrengthGreaterThan(int strength) {
         return heroRepository.findByStrengthGreaterThan(strength);
